# docs/openapi/aether-v0.yaml
openapi: 3.1.0
info:
  title: Aether /api/v0 API
  version: 0.1.0
  contact:
    name: Andres Bott
    url: https://github.com/andresbott/aether
  description: >
    Internal, admin-only server-management API — collections/libraries,
    scanning, tasks/scheduling, and other administration concerns. This is
    NOT the music-client API: browsing, playback, playlists, album/artist
    navigation and search live exclusively under `/rest`, which must stay
    OpenSubsonic-compliant and has no OpenAPI document of its own (a Subsonic
    client does not consume one).


    **Scope.** This document describes the full `/api/v0` surface as mounted
    in `app/router/api_v0.go`: the metadata editor (pictures, raw tags,
    folder/track browsing, identify and identify-album, artist images),
    artists, users, tasks, tokens, radiobrowser, auth, libraries, and
    health/version/me.
servers:
  - url: /api/v0
    description: >
      Every path below is mount-relative. The base is modelled exclusively
      through this `servers` entry, so bumping the version prefix — this is
      `/api/v0` while the API is pre-1.0 — stays a one-line change, with no
      path or operationId edits.
...
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Password login — verifies credentials and starts a session.
      description: >
        Public. The entire request/response contract here is owned by the
        external go-bumbu login library (github.com/go-bumbu/userauth,
        flow/login/handlers.JSON.LoginHandler) wired in
        `handlers/auth/auth.go` over a single-factor password policy — not an
        aether-authored shape. Every credential-shaped failure (unknown user,
        disabled user, wrong password) answers the identical 401, by design,
        so the response can never be used to enumerate accounts. Success sets
        the `sessionCookie` session cookie via `Set-Cookie`; the body itself
        carries no identity.


        Unlike the rest of `/api/v0`, go-bumbu's own error body —
        `{"error": "<message>"}` under plain `application/json` — passes
        through the router's error-envelope middleware untouched, because it
        is already a well-formed JSON object: the middleware only reshapes
        bare plain-text errors into RFC 9457 `application/problem+json`, and
        forwards a body that is already valid JSON as-is (see
        `app/router/errors.go`). So login's errors are documented here with
        their own `AuthError` schema, not the shared `BadRequest`/
        `Unauthorized` responses used elsewhere, which really are
        `problem+json`.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: >
            Login accepted. `done: true` in every case reachable today —
            aether wires only a single password factor; `next` exists for
            multi-factor policies the library supports but aether does not
            yet use.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        '400':
          description: Malformed JSON body, or a missing username/password.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
        '401':
          description: >
            Uniform credential failure: unknown user, disabled user and wrong
            password are all indistinguishable from one another.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
  /auth/logout:
    post:
      operationId: logout
      summary: Clear the session cookie.
      description: >
        Public and idempotent: answers 200 whether or not the caller actually
        held a session. Owned by go-bumbu's `cookieauth.LogoutHandler`
        (`handlers/auth/auth.go`), wrapped by one aether-side addition: an
        optional `{tokenId}` body names the caller's own mintToken session
        token so the server can best-effort revoke it too. A missing or
        malformed body is tolerated silently and never fails the logout; the
        revoke itself is also best-effort and never surfaces a failure.
      tags: [auth]
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '200':
          description: Session cleared (or already absent). No body.
  /auth/password:
    put:
      operationId: changePassword
      summary: Change your own password.
      description: >
        Any authenticated native user (not admin-gated: it acts only on the
        caller's own credential). The current password is re-verified before
        the change — a live session is not by itself authority to change the
        credential that mints it — with per-user brute-force backoff, so a
        stolen session cannot turn this into a password oracle; repeated wrong
        guesses answer 429 with `Retry-After`. On success the caller's own
        session cookie is cleared, so this device is signed out and the user
        re-authenticates with the new password. Native mode only — under
        proxy-header the IdP owns credentials and this route is unmounted.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204':
          description: >
            Password changed. The `sessionCookie` is cleared via `Set-Cookie`
            (the caller is signed out); no body.
        '400':
          description: Malformed JSON, a missing current password, or an empty new password.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: >
            The current password is incorrect. Deliberately 403, not 401: the
            session is still valid — only the re-auth check failed — so a
            client must show a form error, not sign the user out. A genuine
            lost session is the 401 above.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /auth/token:
    post:
      operationId: mintToken
      summary: Mint the SPA's short-lived, per-device session token.
      description: >
        Any authenticated user; not admin-gated (operates only on the
        caller's own tokens). Mints a spa-scoped, 48h-lived token — the
        device-bound counterpart to the session cookie that `/rest` clients
        use to act as the logged-in SPA. Before minting, it sweeps the
        caller's own spa-scoped tokens: every expired one, and the live one
        already bound to this same `deviceId` (so repeated boots of one app
        instance supersede rather than pile up); it also evicts the
        least-recently-used session past a 10-per-user ceiling. Other
        devices' live sessions are left alone, which is what lets the same
        user stay signed in from several first-party apps at once.
      tags: [tokens]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MintTokenRequest' }
      responses:
        '201':
          description: Minted. The plaintext `token` appears here and nowhere else.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MintTokenResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /auth/tokens:
    get:
      operationId: listTokens
      summary: The caller's tokens — user-created PATs and live first-party sessions.
      description: >
        Any authenticated user; not admin-gated (operates only on the
        caller's own tokens). Reports user-created PATs (`kind: "client"`)
        together with the caller's live mintToken sessions (`kind:
        "session"`) so the settings UI can list and tell them apart; an
        expired session token is dropped rather than listed, since it is
        already-superseded plumbing the user should not see.
      tags: [tokens]
      responses:
        '200':
          description: The caller's tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createToken
      summary: Create a user-managed PAT (kind "client") for a third-party client.
      description: >
        Any authenticated user; not admin-gated (operates only on the
        caller's own tokens). `type` picks the credential shape a
        third-party Subsonic client presents: "apikey" (default) is
        hash-only and presented whole via `apiKey=`; "usertoken" is a
        recoverable secret presented as a virtual username + password pair
        for `u`+`t`+`s`/`u`+`p` clients. The plaintext `token` — and for
        "usertoken", also `username`/`password` — appear in this response
        and nowhere else.
      tags: [tokens]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTokenRequest' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateTokenResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '501':
          description: >
            `type: "usertoken"` was requested but the server has no cipher
            configured to store a recoverable secret (`pat.ErrNoCipher`);
            "apikey" tokens are unaffected.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /tasks:
    get:
      operationId: listTasks
      summary: List every registered task, with its schedule if one is configured.
      description: >
        Admin. `apptasks.AvailableTasks` (app/tasks/tasks.go) enumerates the
        fixed set of tasks aether ships today (scan, scan-full,
        fetch-artist-images) — not a user-extensible registry, so there is
        no create/delete for the task definitions themselves, only for their
        schedules.
      tags: [tasks]
      responses:
        '200':
          description: Every task definition, joined with its schedule where one exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /tasks/{name}:
    get:
      operationId: getTask
      summary: Fetch one task definition and its schedule, if any.
      description: >
        Admin. Bare `Task` — unlike `listTasks`, the response is not wrapped
        in an envelope object. See the `Task` schema for the anonymous-embed
        flattening (`TaskWithSchedule` embeds `apptasks.TaskDef`) this
        response relies on.
      tags: [tasks]
      parameters:
        - name: name
          in: path
          required: true
          description: >
            The task's stable id (e.g. "scan"), not its display name.
            Deliberately not the shared `ResourceId` parameter — task ids
            are opaque strings, never numeric.
          schema: { type: string }
      responses:
        '200':
          description: The task and its schedule, if one is configured.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: upsertTaskSchedule
      summary: Create or fully replace a task's schedule.
      description: >
        Admin. Bare `Task`, like `getTask` — not wrapped. Creates the
        schedule when the task has none yet, otherwise replaces it outright.
        `cron_expression` is required (blank answers 400); a 5-field value
        is normalized by prepending a "0" seconds field. `enabled` omitted
        defaults to true — a plain default, not "leave unchanged" (contrast
        `patchTaskSchedule`, which is genuinely tri-state). Answers 404 for
        an unknown task id, and 503 when the schedule store is unavailable.
      tags: [tasks]
      parameters:
        - name: name
          in: path
          required: true
          description: The task's stable id. See `getTask`.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertTaskScheduleRequest' }
      responses:
        '200':
          description: The task with its newly-saved schedule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      operationId: patchTaskSchedule
      summary: Partially update an existing task's schedule.
      description: >
        Admin. Bare `Task`, like `getTask`. Genuinely tri-state: both
        `cron_expression` and `enabled` are optional, and omitting either
        (or sending it explicit null) leaves that attribute unchanged — see
        `PatchTaskScheduleRequest`. Requires a schedule to already exist:
        answers 404 both for an unknown task id and for a known task with no
        schedule yet (create one first with `upsertTaskSchedule`); the two
        causes are not distinguished. Answers 503 when the schedule store is
        unavailable.
      tags: [tasks]
      parameters:
        - name: name
          in: path
          required: true
          description: The task's stable id. See `getTask`.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchTaskScheduleRequest' }
      responses:
        '200':
          description: The task with its updated schedule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      operationId: deleteTaskSchedule
      summary: Remove a task's schedule (the task definition itself is never deletable).
      description: >
        Admin. Answers 404 both for an unknown task id and for a known task
        with no schedule to delete — the two causes are not distinguished,
        like `patchTaskSchedule`. Answers 503 when the schedule store is
        unavailable.
      tags: [tasks]
      parameters:
        - name: name
          in: path
          required: true
          description: The task's stable id. See `getTask`.
          schema: { type: string }
      responses:
        '204':
          description: Schedule deleted. No body.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{name}/trigger:
    post:
      operationId: triggerTask
      summary: Enqueue an ad hoc run of a task, independent of its schedule.
      description: >
        Admin. Works whether or not the task has a schedule configured — a
        manual "run now", unrelated to `upsertTaskSchedule`. Answers 404 for
        an unknown task id. Answers 429 when the run queue is full
        (`taskrunner.ErrQueueFull`): this is the one operation on this
        surface where the handler calls `httperr.Write` directly instead of
        a bare `http.Error`, and it uses its own `queue_full` Problem type
        slug rather than the router-level error envelope's generic
        `rate_limited` slug a plain 429 would otherwise be given — see the
        429 response below.
      tags: [tasks]
      parameters:
        - name: name
          in: path
          required: true
          description: The task's stable id. See `getTask`.
          schema: { type: string }
      responses:
        '202':
          description: Enqueued; the run has not necessarily started yet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TriggerTaskResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429':
          description: >
            The run queue is full. Unlike every other Problem on this
            surface, `type` here is
            "https://aether.local/probs/queue_full" — a handler-authored
            slug, not the generic `rate_limited` a bare `http.Error(w, ...,
            429)` would receive from the router's error envelope.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /tasks/executions:
    get:
      operationId: listTaskExecutions
      summary: Task run history — queued, running, and finished executions.
      description: >
        Admin. Bounded by the runner's history size
        (`taskrunner.Cfg.HistorySize`, default 20) — not a paginated,
        unbounded log. Ordered newest-queued first.
      tags: [tasks]
      responses:
        '200':
          description: The run history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskExecutionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/executions/{id}/cancel:
    post:
      operationId: cancelTaskExecution
      summary: Cancel a queued or running task execution.
      description: >
        Admin. A waiting execution is canceled outright; a running one is
        asked to stop via context cancellation. Answers 409 — not 404 — for
        an execution id that does not exist at all: the handler maps every
        failure from the underlying cancel call (unknown id, or an
        already-terminal execution that cannot be canceled) to the same
        Conflict response, without distinguishing them.
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The execution id, as returned by `triggerTask`'s
            `execution_id` or `listTaskExecutions`' `id`.
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Canceled (or cancellation requested). No body.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/executions/{id}/logs:
    get:
      operationId: getTaskExecutionLog
      summary: The captured log output of one task execution.
      description: >
        Admin. `text/plain`, not JSON — the one response on this surface
        that isn't `application/json`/`application/problem+json`. Answers
        503 when no log sink is configured for this deployment
        (`TaskLogGetter` unset). A read failure for an id that does not
        exist, or whose log file is no longer on disk, answers 500 rather
        than 404 — the handler does not distinguish the two — and, like
        every other 500 on this surface, is not itemized as a documented
        response here.
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The execution id, as returned by `triggerTask`'s
            `execution_id` or `listTaskExecutions`' `id`.
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The execution's captured log output.
          content:
            text/plain:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: _aether_login
      description: >
        Native-mode session cookie (see docs/agents/authentication.md).
        `/api/v0` defaults every route to admin-only, so a valid, enabled
        admin session is required for the whole metadata surface. This is the
        only credential `/api/v0` actually validates: no route on this
        surface reads an `Authorization` header, and Aether's PATs
        (bearer-shaped credentials) are verified exclusively on `/rest` (via
        `apiKey`, or `u`+`t`+`s` / `u`+`p`), never here.


        In **proxy-header mode**, this cookie plays no role either — identity
        instead comes from reverse-proxy-injected headers (`Remote-User` /
        `Remote-Groups`), validated by `headerGuard`. That trust is
        established between the proxy and Aether, not chosen by the API
        caller, so it has no client-facing credential to model as an OpenAPI
        security scheme; it is intentionally left undeclared here rather than
        represented by a scheme nothing in the protocol actually sends.
...
  schemas:
...
    LoginRequest:
      type: object
      description: >
        Owned by the external go-bumbu login library
        (flow/login/handlers.JSON.LoginPayload), not an aether type — see
        `login`'s description.
      properties:
        username: { type: string }
        password: { type: string }
        sessionRenew:
          type: boolean
          default: false
          description: >
            "Remember me": opts the session into rolling expiry renewal
            (activity keeps it alive) instead of a fixed window.
      required: [username, password]
...